refactor PostTypeLabels to extend new base class, and move Labels to new namespace
