+++ src/Post/PostTypeLabels.php
@@ -2,10 +2,9 @@
 
 namespace Silk\Post;
 
-use Illuminate\Support\Collection;
-use Silk\Support\LabelsCollection;
+use Silk\Labels\Labels;
 
-class PostTypeLabels
+class PostTypeLabels extends Labels
 {
     /**
      * Labels referencing the singular form
@@ -38,66 +37,4 @@ class PostTypeLabels
         'not_found'             => 'No %s found.',
         'search_items'          => 'Search %s',
     ];
-
-    /**
-     * The master collection of labels
-     * @var Collection
-     */
-    protected $labels;
-
-    /**
-     * Set the singular labels using the given form.
-     *
-     * @param $label The singular label form to use
-     *
-     * @return $this
-     */
-    public function setSingular($label)
-    {
-        $this->merge(
-            LabelsCollection::make($this->singular)->setForm($label)
-        );
-
-        return $this;
-    }
-
-    /**
-     * Set the plural labels using the given form.
-     *
-     * @param $label The plural label form to use
-     *
-     * @return $this
-     */
-    public function setPlural($label)
-    {
-        $this->merge(
-            LabelsCollection::make($this->plural)->setForm($label)
-        );
-
-        return $this;
-    }
-
-    /**
-     * Get all the labels as an array.
-     *
-     * @return array
-     */
-    public function toArray()
-    {
-        return $this->labels ? $this->labels->toArray() : [];
-    }
-
-    /**
-     * Merge the labels with the master collection.
-     *
-     * @param  LabelsCollection $collection
-     */
-    protected function merge(LabelsCollection $collection)
-    {
-        if (! $this->labels) {
-            $this->labels = new Collection;
-        }
-
-        $this->labels = $this->labels->merge($collection->replaced());
-    }
 }
